fix(callback): replace worker with handler
